Moved teamPage folder down
