+++ main.py
@@ -3,7 +3,7 @@ import streamlit as st
 
 st.title("Competitie Planner BC Geldrop")
 
-homePage = st.Page("pages/mainPage.py", title="Uitleg")
+homePage = st.Page("pages/src/mainPage.py", title="Uitleg")
 
 # Create a list of team names
 team1 = st.Page("pages/BCGeldrop1.py", title="BC Geldrop 1")
